202411：版本2.1.5   修改bf_aio库及bfcgate，在广播发送时针对订阅和发送一起加锁，防止出现行情时间戳乱序的问题，修改bfsap写日志的bug

+++ src/base/kernel/BF_Tools.cpp
@@ -879,7 +879,7 @@ int CBF_Tools::RemoveOverdueFile(const char *delpath,const char *fileattr, int o
    
 #ifdef _WINDOWS
 	struct _finddata_t fd;
-	int dh;
+	intptr_t dh;
 	dh = _findfirst(szfile,&fd);
 	if (dh ==-1 ) 
 	{
